Getting a simple graphql test going.
In order to use `factory.create` I've written a simple rethinkdb adapter
for factory-girl. If this works we should open source that, but the
interface is a little strange since there's no ORM model. Need to
think about weather that's the best way to do this.

The tests run in a test database that you'll have to create and migrate
separate from your development database. I added instructions for doing
that to the README. We might be able to make that a little simpler.

This commit breaks the data generation scripts since they relied on the
factories producing data the way it looks in the api and not the way it
looks in the database. We'll need to create some new factories for them
or change the script to use the new factories before this gets merged.