adding cycleLaunched change feed. Refactoring configureChangeFeeds
Now that we have 3 of these it seemed like a good time to refactor them
into individial files. I also changes the api a little so that the queue
is injected into the setup function which made testing easier.

I'm building my own mock in this test but we could easily move to a mock
library, just didn't feel like doing the research into the best option
just yet.