handle both users with the 'player' and 'moderator' roles
This worker now (somewhat) gracefully handles duplicate users in the queue. It
does a `.replace()` in RethinkDB rather than an `.insert()`. This doesn't
prevent the GitHub API call from getting executed multiple times, but the only
harm there is ... multiple invocations to the API that won't hurt anything
(from a data integrity perspective).