Code review fixes
Lots of spelling, some better variable names. A few logic errors in the
tests.

Also making `subject` in the SurveyQuestionItem an array, which makes it
match the first use case, but won't work for other types of single
subject questions. Like Response and InputResponse, we'll have to do
something more robust to handle multiple types. I'll work on that for
the next PR.