if GitHub team already exists, fetch it
This should really only happen if a chapter with the same channelName gets created more than once. In reality, this should only happen when running the `npm run data:playtest` script.

We will inspect the response body that comes back from GitHub to see if it is this one particular error code. If it is, we'll call back out to GitHub to fetch the correct team with the same name (so that we can update the githubTeamId on the chapter).