refactoring retro completion notifications
Instead of having a bunch of logic in the changefeed config to figure
out if a given response represents that last response for that
respondant to a survey, I've changed the changefeed to just push
messages on every response submission. The worker is now doing the work
of determining if the response represents the completion of a retro
survey.

There are a few benefits to this, but the largest is that worker code
has tests and change feeds are really hard to test due to race
conditions.