fix survey creation bug
When creating surveys for a cycle we were pulling all projects for the
chapter, not just the ones active in this cycle, which was causing
errors. Once I looked into that I noticed that we were making a similar
mistake when notifying project rooms about the the starts of cycle
reflection, we should only notify projects for that cycle, not every
project that has ever been active in that chapter.