adding updateTeamECCStats action and calling it on survey completion
This updates `player.ecc` as well as `player.cycleProjectECC`. The
latter is a nested map that stores the ecc gained for any given project
in a cycle at `player[cycleId][projectId]`.

This serves 2 purposes:

1. If we even need to change the way we aggregate ecc we still have the raw
values.

2. We can make the update idempotent by checking so see if we've already
recorded the ecc for a particular project this cycle and, if so, just
recalculating the aggregate ecc with the new value instead of adding ecc
for a single project twice. This is needed becuase players can change
their answers to a survey even after they have completed all the
questions, so this may get run multiple times for the same project.