Fixing a few bugs in the newOrUpdatedVote worker.
Fixes #296

**The Bugs:**

1. newOrUpdatedVote.js was invoking the cycleVotingResults graphql API
and using `true` as the value of `currentUser`. That worked find when
the API was just checking for the presence of that attribute, but now
that the API actually tries to look up the user in the players or
moderators table it breaks.

2. `pushCandidateGoalsForCycle()` created a promise but didn't return
it, so if the promise got rejected (as was happening because of #1)
that error got ignored.

**The Fix:**

The non-graphql-specific logic for the getCycleVotingResults API has
been moved into an action so that it can be reused by the
newOrUpdatedVote worker. The worker is also not explicitly awaiting the
`pushCandidateGoalsForCycle()` promise so we'll get notified of errors.