`graphql/models/Player` --> `graphql/models/User`
Eventually, it's likely that we will merge the `players` and `moderators` table. Already, we need an API endpoint to `getUserById`, whether the user is a player or a moderator. This commit renames the `Player` folder to `User` so that we have a sane place to put the endpoint, but does **not** go as far as doing the refactor that merges the two tables. Baby steps.