adding an hours question to the survey
It occured to me that the "how many hours" question could be implemented
as a question about the project, not about the respondent. This turned
out be be pretty simple. Since we don't have integer questions yet I
just made it a text response. We can change that as soon as the UI
supports numeric inputs.

We still need to change how we do surveys to support self questions and
delivering different questions to different users, but this will allow
us to start collecting data.