adding explicit weights to objectives.
Fixes #507

Up to this point we've been weighting each objective as exponentially
less important than the one before it in the list, this way we can make
some objectives the same weight.

I'm also adding code to pull the objectives config out of the pool,
which, in addition to helping testing, will allow us to experiment more
once the moderator can manually configure the pools.