fixing 3 bugs that prevented TFA from considering player feedback right
This fixes three bugs:

1. The keys in the object returned by `getLatestFeedbackStats` were
camelCase, but we were expecting UPPER_CASE in
`PlayersGetTeammatesTheyGaveGoodFeedbackAppraiser`. Changed to
camelCase.

2. The values returned by getLatestFeedbackStats were the raw likert
responses, but `PlayersGetTeammatesTheyGaveGoodFeedbackAppraiser` was
expecting the "stat" in the 0-100 scale. Changed to 0-100 scale

3. Likert responses were being handled wrong, with `0` treated as "the
worst possible rating" instead of "n/a" or "not enough information" and
`1` treated like "the next to worst rating" instead of the worst, etc.
I've hooked into the `likert7Average` method to convert the 1-7
responses to the appropriate percentage, and I'm treating an "n/a"
response the same as if there were no feedback at all, which in this
case will cause the scoring function to rate that pairing very high as
if you've never worked with that person before.