refactor `loadPlayers` action to reduce "magic"
Previously, the `loadPlayers` action would load the players, then query IDM and load the users, then merge the objects together. Instead, we are now keeping both `players` and `users` in the state, which makes the code much cleaner and easier to understand.

This was partially because we're also going to want `users` in the state when we load voting results.

In addition, we improved the players view by displaying chapter name rather than id.