allow: skip channel creation, override project name
Modify import script so that you can pass a command-line argument that
will cause channel creation to be skipped when a new project is created.

You can now also pass-in the input file.

Lastly, project name can be supplied along with the goal number if we
want to force projects to be named a certain way.

Side-note: this is all to support the case where channels were created
but the corresponding database saves failed.