on pool split, if players have no stats at al treat them as if elo was 0
This is an edge case that probably means something went wrong somewhere
else in the code, but if players are voting that have no ELO at all, we
need to figure out how to sort them when splitting the pool. With this
change I'm sorting them all to the bottom as if they had an ELO of 0.