update project formation readme.
One of the queries needed to be updated to work with votes in pools, not
cycles. I also changed the instructions to use the new resyncdb script and to
use a different database name to reduce the risk of someobe accidentally
running a command in the wrong window and writing to the production
database.