never collapse the voting results if there's just one pool
when there's only one pool there's no way to uncollapse the results, so
if you're not in the pool it starts off collapsed and you have no way to
open it. The immediate reason I need this fixed is that when we migrate
old cycles to pools it only places people who voted in the pool, so I
can't load the cycle-voting-results page and see anything useful becuase
I didn't vote.