Fixing two bugs:
* normalizr expects an id in the cycleVotingResults. We rere only adding
it after a graphQL call, not when we get it via a socket. Now we're
adding it in the server action. Eventually we'd like this to just be the
cycleId
* We were processing cycleVotingResults socket messages multiple times
as watchers got left on the socket. Calling unwatch seems to fix that.
We also cleaned up the sub/unsub logic a bit.