load only as-yet-unloaded users when new voting results come in
We really only need to load the users from IDM when the voting results first load. It's extremely unlikely that the user data would change between votes. This change has the nice side-effect that, since `users` in the state won't be busy, the voting UI won't flash every time a new vote comes in.