Fixing nocking
We were nocking 2 IDM calls before all tests but didn;t really need to
be. In fact, in the last test we needed to mock something else
altogether, so we were calling nockClean() and then re-nocking the IDM
calls with different values. With this change I'm just removing the
beforeEach nocking and instead just making sure that any nocks from
previous tests are cleared in the beforeEach. Then I'm adding the needed
nock calls to the one test that needs them. This test actually needed to
support multiple calls to the same IDM API, so I added that
functionality to the fixtures.