removing getMinTeamSize
This concept of minTeamSize is actually a little more nuanced now. It
used to be that no team could be smaller than 2 players. Now a team can
be 1-player if the recommended goal size is 1 player, but otherwise must
be 2 or greater. Becuase of that there isn't a single "minimum team
size" so I'm removing that helper function in favor of context specific
numbers.