fixing project closing bug
There were 2 bugs in the project closing logic (where project stats
and RXP and review accuracy are calulated) that were causing errors.

1. We were incorrectly assuming that all closed projects would have a
quality and completeness stat. That's not true since there are a lot of
legacy projects that we did not do those caluclations for since they
predated the RXP stats that gave project reviews more meaning.

2. In the surveyService we were incorrectly assuming that all questions
about a project would be associated with a stat in
getStatResponsesBySubjectId. That's not true and we need to make sure to
filter out responses are not for stats-related questions.