pull coaches from users in state in UI instead of asking from query
The way we were asking for coaches was causing an n+1 query situation
where on the server we were spawning over 100 getUserById queries to
idm every time the project page loaded. Now we just get the ids from
game and pull the users out of the state.

Related to [ch1912]