new XP formula based on project completeness
1. Changing the forumula for XP to depend on project completeness and a
`xpValue` and `dynamic` attributes of the goal. This means that we have
to update player stats for a project when the project closes. We also
have to migrate old goal data in votes and projects to include the new
`xpValue` and `dynamic` attributes.

2. adding a script to backfill project reviews for recent
projects that never got an external review. We want to do this to make
sure that people have as many projects with completeness scores in their
recent history as possible, and since coaches were not doing as many
reviews a few weeks ago we're just going to trust the internal
completeness scores and add an external review that's just the average
of the internal ones.

3. changing the signature of the buildSurvey fixture since it has so many
optional arguments

4. better logging and error handling around project initialization. This is
mainly to make things smoother in development where slack errors about
users not existing and the like are mor common, but either way we don't
want slack errors to scuttle the whole initialization process

fix tests