scaling project completeness based on project hours
The new XP score is build on both completeness and contribution, but we
run into a problem with those metrics becuase they currently assume that
everyone is working all week every week. This leads to the following
problems.

The guild closes some days, so some weeks there are not 38 hours for
working on a project, but the projects are designed for 38 hours of work
and the completness metric assumes that. So teams will likely complete
about 20% less on their projects in a short week, but they should not be
penalized for that.

Even in a normal week, if my teammate takes a personal day it means we
have fewer resources as a team and so the project completeness will be
lower. I shouldn't take an XP hit just becuase my teammate took a
personal day.

In addition to the impact on project completeness, when I take a
personal day and am out 20% of the week I will (correctly) get lower
contribution scores from my team. Those lower scores will translate into
a XP penalty just for taking one of my alloted days off.

The fix we're going to employ is to scale completeness and contribution
by multiplying them by `(expectedProjectHours / actualProjectHours)`

This commit adds the scaling to completeness. The "raw" completness is
still stored as `rawProjectCompleteness` and we still use that to rate
players review accuracy.